requirements.txt
# File containing requirements for running the deployer

pyyaml
kubernetes
